@@ -40,41 +40,47 @@
     "tag": "next"
   },
   "dependencies": {
-    "@azure/ms-rest-nodeauth": "^2.0.2",
-    "@types/node": "^12.6.9",
-    "big-number": "1.0.0",
+    "@azure/ms-rest-nodeauth": "2.0.2",
+    "@js-joda/core": "^2.0.0",
     "bl": "^3.0.0",
     "depd": "^2.0.0",
-    "iconv-lite": "^0.4.23",
+    "iconv-lite": "^0.5.0",
+    "jsbi": "^3.1.1",
     "native-duplexpair": "^1.0.0",
     "punycode": "^2.1.0",
-    "readable-stream": "^3.4.0",
+    "readable-stream": "^3.5.0",
     "sprintf-js": "^1.1.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.5.5",
-    "@babel/node": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-typescript": "^7.3.3",
-    "@babel/register": "^7.5.5",
-    "@commitlint/cli": "^8.0.0",
-    "@commitlint/config-conventional": "^8.0.0",
-    "@commitlint/travis-cli": "^8.0.0",
-    "async": "^2.6.1",
-    "babel-eslint": "^11.0.0-beta.0",
+    "@babel/cli": "^7.8.3",
+    "@babel/core": "^7.8.3",
+    "@babel/node": "^7.8.3",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/preset-env": "^7.8.3",
+    "@babel/preset-typescript": "^7.8.3",
+    "@babel/register": "^7.8.3",
+    "@commitlint/cli": "^8.3.5",
+    "@commitlint/config-conventional": "^8.3.4",
+    "@commitlint/travis-cli": "^8.3.5",
+    "@types/bl": "^2.1.0",
+    "@types/depd": "^1.1.32",
+    "@types/node": "^13.5.1",
+    "@types/readable-stream": "^2.3.5",
+    "@types/sprintf-js": "^1.1.2",
+    "@typescript-eslint/eslint-plugin": "^2.18.0",
+    "@typescript-eslint/parser": "^2.18.0",
+    "async": "^2.6.3",
     "babel-plugin-istanbul": "^5.2.0",
-    "benchmark": "^2.1.0",
     "chai": "^4.1.2",
-    "codecov": "^3.5.0",
+    "codecov": "^3.6.1",
     "eslint": "^5.16.0",
     "mitm": "^1.7.0",
-    "mocha": "^6.1.4",
+    "mocha": "^6.2.2",
     "nyc": "^14.1.1",
-    "rimraf": "^2.6.3",
-    "semantic-release": "^15.13.12",
-    "sinon": "^7.3.2",
-    "typescript": "^3.6.2"
+    "rimraf": "^3.0.0",
+    "semantic-release": "^17.0.1",
+    "sinon": "^7.5.0",
+    "typescript": "^3.7.5"
   },
   "scripts": {
     "lint": "eslint src test --ext .js,.ts && tsc",
@@ -98,7 +104,26 @@
           }
         }
       ],
-      "@babel/preset-typescript"
+      [
+        "@babel/preset-typescript",
+        {
+          "allowDeclareFields": true
+        }
+      ]
+    ],
+    "plugins": [
+      [
+        "@babel/transform-typescript",
+        {
+          "allowDeclareFields": true
+        }
+      ],
+      [
+        "@babel/plugin-proposal-class-properties",
+        {
+          "loose": true
+        }
+      ]
     ]
   },
   "commitlint": {
