@@ -1,4 +1,4 @@
-import { numberToInt64LE } from './bigint';
+import JSBI from 'jsbi';
 
 const SHIFT_LEFT_32 = (1 << 16) * (1 << 16);
 const SHIFT_RIGHT_32 = 1 / SHIFT_LEFT_32;
@@ -106,9 +106,36 @@ class WritableTrackingBuffer {
     this.position += length;
   }
 
+  writeBigInt64LE(value: JSBI) {
+    this.writeBigU_Int64LE(value);
+  }
+
+  private writeBigU_Int64LE(value: JSBI) {
+    this.makeRoomFor(8);
+
+    let lo = JSBI.toNumber(JSBI.bitwiseAnd(value, JSBI.BigInt(0xffffffff)));
+
+    this.buffer[this.position++] = lo;
+    lo = lo >> 8;
+    this.buffer[this.position++] = lo;
+    lo = lo >> 8;
+    this.buffer[this.position++] = lo;
+    lo = lo >> 8;
+    this.buffer[this.position++] = lo;
+
+    let hi = JSBI.toNumber(JSBI.bitwiseAnd(JSBI.signedRightShift(value, JSBI.BigInt(32)), JSBI.BigInt(0xffffffff)));
+
+    this.buffer[this.position++] = hi;
+    hi = hi >> 8;
+    this.buffer[this.position++] = hi;
+    hi = hi >> 8;
+    this.buffer[this.position++] = hi;
+    hi = hi >> 8;
+    this.buffer[this.position++] = hi;
+  }
+
   writeInt64LE(value: number) {
-    const buf = numberToInt64LE(value);
-    this.copyFrom(buf);
+    this.writeBigInt64LE(JSBI.BigInt(value));
   }
 
   writeUInt32BE(value: number) {
@@ -125,8 +152,11 @@ class WritableTrackingBuffer {
   }
 
   writeUInt64LE(value: number) {
-    this.writeInt32LE(value & -1);
-    this.writeUInt32LE(Math.floor(value * SHIFT_RIGHT_32));
+    this.writeBigUInt64LE(JSBI.BigInt(value));
+  }
+
+  writeBigUInt64LE(value: JSBI) {
+    this.writeBigU_Int64LE(value);
   }
 
   writeInt8(value: number) {
