@@ -167,17 +167,17 @@ describe('Guid Parser Test', function() {
     }
   });
 
-  it('arrayToLowerCaseGuid', () => {
+  it('bufferToLowerCaseGuid', () => {
     for (let i = 0; i < arrays.length; i++) {
       const array = arrays[i];
-      assert.strictEqual(guidParser.arrayToLowerCaseGuid(array), lowerGuids[i]);
+      assert.strictEqual(guidParser.bufferToLowerCaseGuid(array), lowerGuids[i]);
     }
   });
 
-  it('arrayToUpperCaseGuid', () => {
+  it('bufferToUpperCaseGuid', () => {
     for (let i = 0; i < arrays.length; i++) {
       const array = arrays[i];
-      assert.strictEqual(guidParser.arrayToUpperCaseGuid(array), upperGuids[i]);
+      assert.strictEqual(guidParser.bufferToUpperCaseGuid(array), upperGuids[i]);
     }
   });
 });
